Extract CSV parser to separate file
