+++ lib/command/licenses_command.dart
@@ -1,5 +1,3 @@
-import 'dart:io';
-
 import '../service/bom_service.dart';
 import 'abstract_command.dart';
 
@@ -19,8 +17,8 @@ class LicensesCommand extends AbstractCommand {
   Future<void> execute() async {
     final result =
         await service.compareLicenses(licensesFile: file, diffOnly: diffOnly);
-    stdout.writeln('Shared BOM size: ${result.bom}');
-    stdout.writeln('Matching licenses: ${result.common}');
-    stdout.writeln('Different licenses: ${result.bom - result.common}');
+    print('Shared BOM size: ${result.bom}');
+    print('Matching licenses: ${result.common}');
+    print('Different licenses: ${result.bom - result.common}');
   }
 }
